requirements-prod.txt
Flask==2.3.3
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
gunicorn==21.2.0
flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
flask-admin==1.6.1
flask-login==0.6.3
Flask-JWT-Extended==4.5.3
PyJWT==2.8.0
flask-migrate==4.0.5
requests==2.31.0
Flask-Caching==2.1.0
Flask-Limiter==3.5.0
redis==5.0.1
